Move files in received tmp folder to error folder if not recongnized
When finding invalid files in tmp folder, move them to error folder.
This situation can only happen if files are manually copied to tmp folder.

The receiver process uses a timestamp on each received file. The parsing of
this timestamp is used to know whether a tmp file is considered valid.

The parsing of tmp files ensures that processed files (in processed folder)
can be processed again if they are moved to tmp folder.