Clarify code by removing implicit use of Maybe
Removed use of 'do' notation in this function. Although it works, it's not
clear that it short-circuits the entire function if a date cannot be parsed,
which means the last line is never run.

By using let and applicative the intent is much more clear and Just/Nothing
both are clearly expected